+++ src/components/PluginTrends/Charts/PluginInstallationsChart.tsx
@@ -1,31 +1,25 @@
-import React, { useEffect, useRef, useMemo } from 'react'
+import React, { useEffect, useRef, useMemo, useState } from 'react'
 import * as echarts from 'echarts'
-import { PluginChartProps } from '../../data/plugins'
 import dayjs from 'dayjs'
+import { PluginChartProps } from '../../../data/plugins'
+import ResetZoomButton from './ResetZoomButton'
 
-const PluginInstallationsPerVersion: React.FC<PluginChartProps> = ({ data }) => {
-    const chartRef = useRef(null)
+const PluginInstallationsChart: React.FC<PluginChartProps> = ({ data }) => {
+    const chartRef = useRef<HTMLDivElement | null>(null)
+    const [chartInstance, setChartInstance] = useState<echarts.ECharts | null>(null) // State to hold the chart instance
 
     const formattedData = useMemo(() => {
-        if (!data || !data.installationsPerVersion) return []
-        return Object.entries(data.installationsPerVersion).map(([version, installations]) => ({
-            version,
+        if (!data || !data.installations) return []
+        return Object.entries(data.installations).map(([timestamp, installations]) => ({
+            date: dayjs(parseInt(timestamp)).format('MMM YYYY'),
             installations,
         }))
-    }, [data])
-
-    const formattedDate = useMemo(() => {
-        if (!data || !data.installations) return ''
-        const latestDate = Object.entries(data.installations).reduce((acc, [timestamp]) => {
-            return parseInt(timestamp) > acc ? parseInt(timestamp) : acc
-        }, 0)
-        return dayjs(latestDate).format('MMM YYYY')
     }, [data])
 
     const option = useMemo(() => {
         return {
             title: {
-                text: `Installations Per Version (${formattedDate})`,
+                text: 'Monthly Installations Over Time',
                 left: 'center',
                 textStyle: { fontSize: 16, fontWeight: 'bold' },
             },
@@ -39,15 +33,15 @@ const PluginInstallationsPerVersion: React.FC<PluginChartProps> = ({ data }) => 
                     color: '#fff',
                 },
                 axisPointer: {
-                    type: 'shadow',
-                    shadowStyle: {
-                        color: 'rgba(150, 150, 150, 0.3)',
+                    type: 'line',
+                    lineStyle: {
+                        color: '#777',
                     },
                 },
             },
             xAxis: {
                 type: 'category',
-                data: formattedData.map((item) => item.version),
+                data: formattedData.map((item) => item.date),
                 axisLabel: {
                     fontSize: 12,
                 },
@@ -71,7 +65,7 @@ const PluginInstallationsPerVersion: React.FC<PluginChartProps> = ({ data }) => 
                 axisLabel: {
                     fontSize: 12,
                     formatter: function (value: number) {
-                        return value === 0 ? '' : value // Hide the 0 label
+                        return value === 0 ? '' : `${value / 1000}k` // Hide the 0 label
                     },
                 },
                 axisLine: {
@@ -96,38 +90,56 @@ const PluginInstallationsPerVersion: React.FC<PluginChartProps> = ({ data }) => 
                 bottom: '15%',
                 top: '15%',
             },
+            dataZoom: [
+                {
+                    type: 'inside',
+                    xAxisIndex: 0,
+                    start: 0,
+                    end: 100,
+                },
+            ],
             series: [
                 {
                     data: formattedData.map((item) => item.installations),
-                    type: 'bar',
+                    type: 'line',
+                    smooth: true,
+                    lineStyle: {
+                        width: 2,
+                        color: '#3f51b5',
+                    },
                     itemStyle: {
                         color: '#3f51b5',
                     },
-                    barWidth: '60%',
                 },
             ],
         }
-    }, [formattedData, formattedDate])
+    }, [formattedData])
 
     useEffect(() => {
         if (!chartRef.current) return
 
-        const chart = echarts.init(chartRef.current)
-        chart.setOption(option)
+        const instance = echarts.init(chartRef.current)
+        instance.setOption(option)
+        setChartInstance(instance)
 
         const handleResize = () => {
-            chart.resize()
+            instance.resize()
         }
 
         window.addEventListener('resize', handleResize)
 
         return () => {
             window.removeEventListener('resize', handleResize)
-            chart.dispose()
+            instance.dispose()
         }
     }, [option])
 
-    return <div ref={chartRef} style={{ height: '450px', width: '100%' }} />
+    return (
+        <div>
+            <div ref={chartRef} style={{ height: '450px', width: '100%' }} />
+            <ResetZoomButton chartInstance={chartInstance} />
+        </div>
+    )
 }
 
-export default PluginInstallationsPerVersion
+export default PluginInstallationsChart
