+++ src/components/PluginTrends/Charts/ResetZoomButton.tsx
