+++ src/components/PluginTrends/Layout/PluginCard.tsx
@@ -5,7 +5,7 @@ import { Link } from 'react-router-dom'
 import InsertDriveFileIcon from '@mui/icons-material/InsertDriveFile'
 import ImageIcon from '@mui/icons-material/Image'
 import PluginCardChart from '../Charts/PluginCardChart'
-import { IPluginData } from '../../data/plugins'
+import { IPluginData } from '../../../data/plugins'
 
 interface PluginCardProps {
     plugin: IPluginData
