+++ src/pages/plugin-detail/index.tsx
@@ -5,10 +5,10 @@ import { styled } from '@mui/system'
 
 // import Grid from '@mui/material/Unstable_Grid2'
 import NavBar from '../../components/Layout/NavBar'
-import PluginInstallationsChart from '../../components/Charts/PluginInstallationsChart'
-import PluginInstallationsPerVersion from '../../components/Charts/PluginInstallationsPerVersionChart'
-import PluginInstallationsPercentageChart from '../../components/Charts/PluginInstallationsPercentageChart'
-import PluginInstallationsPercentagePerVersionChart from '../../components/Charts/PluginInstallationsPercentagePerVersionPieChart'
+import PluginInstallationsChart from '../../components/PluginTrends/Charts/PluginInstallationsChart'
+import PluginInstallationsPerVersion from '../../components/PluginTrends/Charts/PluginInstallationsPerVersionChart'
+import PluginInstallationsPercentageChart from '../../components/PluginTrends/Charts/PluginInstallationsPercentageChart'
+import PluginInstallationsPercentagePerVersionChart from '../../components/PluginTrends/Charts/PluginInstallationsPercentagePerVersionPieChart'
 
 import InsertDriveFileIcon from '@mui/icons-material/InsertDriveFile'
 import ImageIcon from '@mui/icons-material/Image'
